+++ src/main/kotlin/kr/hhplus/be/domain/order/OrderEventRequestKafkaPublisher.kt
@@ -1,0 +1,5 @@
+package kr.hhplus.be.domain.order
+
+interface OrderEventRequestKafkaPublisher {
+    fun publish(event: OrderCompletedEvent)
+}
